feat(vs2022_third): C4244 fix double→long conversions in SpecifyCoordinateDlg.cpp — 1 file
Fix 20 C4244 warnings (double→long narrowing conversions) by adding
static_cast<long>() at assignment sites.

No behavior changes, only explicit casts to match target types.

+++ SRC/NanoSpec/SpecifyCoordinateDlg.cpp
@@ -156,9 +156,9 @@ void CSpecifyCoordinateDlg::OnOK()
 	SetOperationLog("OK Button was push.");
 
 	GetDlgItemText(IDC_COORDINATE_X, csBuff);
-	m_lX = atof(csBuff) * MICROMETRE;
+	m_lX = static_cast<long>(atof(csBuff) * MICROMETRE);
 	GetDlgItemText(IDC_COORDINATE_Y, csBuff);
-	m_lY = atof(csBuff) * MICROMETRE;
+	m_lY = static_cast<long>(atof(csBuff) * MICROMETRE);
 	m_iMoveMode = m_iType;
 	CDialog::OnOK();
 }
@@ -189,34 +189,34 @@ BOOL CSpecifyCoordinateDlg::CheckData()
 
 		switch ( m_StageConfig.OrgPos ) {
 		case 1: // 中心
-			coordLeftTop.lX = m_dSampleSizeX / 2 * -1;
-			coordLeftTop.lY = m_dSampleSizeY / 2;
-			coordRightBottom.lX = m_dSampleSizeX / 2;
-			coordRightBottom.lY = m_dSampleSizeY / 2 * -1;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX / 2 * -1);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY / 2);
+			coordRightBottom.lX = static_cast<long>(m_dSampleSizeX / 2);
+			coordRightBottom.lY = static_cast<long>(m_dSampleSizeY / 2 * -1);
 			break;
 		case 2: // 左上
 			coordLeftTop.lX = 0;
 			coordLeftTop.lY = 0;
-			coordRightBottom.lX = m_dSampleSizeX;
-			coordRightBottom.lY = m_dSampleSizeY * -1;
+			coordRightBottom.lX = static_cast<long>(m_dSampleSizeX);
+			coordRightBottom.lY = static_cast<long>(m_dSampleSizeY * -1);
 			break;
 		case 3: // 左下
 			coordLeftTop.lX = 0;
-			coordLeftTop.lY = m_dSampleSizeY;
-			coordRightBottom.lX = m_dSampleSizeX;
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY);
+			coordRightBottom.lX = static_cast<long>(m_dSampleSizeX);
 			coordRightBottom.lY = 0;
 			break;
 		case 4: // 右下
-			coordLeftTop.lX = m_dSampleSizeX * -1;
-			coordLeftTop.lY = m_dSampleSizeY;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX * -1);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY);
 			coordRightBottom.lX = 0;
 			coordRightBottom.lY = 0;
 			break;
 		case 5: // 右上
-			coordLeftTop.lX = m_dSampleSizeX * -1;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX * -1);
 			coordLeftTop.lY = 0;
 			coordRightBottom.lX = 0;
-			coordRightBottom.lY = m_dSampleSizeY * -1;
+			coordRightBottom.lY = static_cast<long>(m_dSampleSizeY * -1);
 			break;
 		}
 
@@ -228,10 +228,10 @@ BOOL CSpecifyCoordinateDlg::CheckData()
 	///// 指定座標が、ステージ稼動範囲であるか
 	case SPECIFY_COORDINATE_STAGE:
 	default:
-		coordLeftTop.lX = m_StageConfig.StageMaxCoord.dLeftX;
-		coordLeftTop.lY = m_StageConfig.StageMaxCoord.dUpY;
-		coordRightBottom.lX = m_StageConfig.StageMaxCoord.dRightX;
-		coordRightBottom.lY = m_StageConfig.StageMaxCoord.dDownY;
+		coordLeftTop.lX = static_cast<long>(m_StageConfig.StageMaxCoord.dLeftX);
+		coordLeftTop.lY = static_cast<long>(m_StageConfig.StageMaxCoord.dUpY);
+		coordRightBottom.lX = static_cast<long>(m_StageConfig.StageMaxCoord.dRightX);
+		coordRightBottom.lY = static_cast<long>(m_StageConfig.StageMaxCoord.dDownY);
 
 		NS_ConvertToStageGetPosCoord(&coordLeftTop);
 		NS_ConvertToStageGetPosCoord(&coordRightBottom);
@@ -245,14 +245,14 @@ BOOL CSpecifyCoordinateDlg::CheckData()
 	if(CheckMinMaxDouble(csBuff, min(coordLeftTop.lX, coordRightBottom.lX) / MICROMETRE, max(coordLeftTop.lX, coordRightBottom.lX) / MICROMETRE) != CHECK_DATA_OK){
 		return FALSE;
 	}
-	lPointX = atof(csBuff) * MICROMETRE;
+	lPointX = static_cast<long>(atof(csBuff) * MICROMETRE);
 
 	///// Y /////
 	GetDlgItemText(IDC_COORDINATE_Y, csBuff);
 	if(CheckMinMaxDouble(csBuff, min(coordLeftTop.lY, coordRightBottom.lY) / MICROMETRE, max(coordLeftTop.lY, coordRightBottom.lY) / MICROMETRE) != CHECK_DATA_OK){
 		return FALSE;
 	}
-	lPointY = atof(csBuff) * MICROMETRE;
+	lPointY = static_cast<long>(atof(csBuff) * MICROMETRE);
 
 	return TRUE;
 }
